refactor: create reusable pattern for process instance diagram overlays
