+++ operate/client/src/App/ProcessInstance/TopPanel/DiagramOverlays/agentStatusOverlay.test.tsx
@@ -7,10 +7,18 @@
  */
 
 import type {AgentInstanceStatus} from '@camunda/camunda-api-zod-schemas/8.10';
-import {AgentStatusOverlay} from './index';
+import {AgentStatusOverlay} from './agentStatusOverlay';
 import {render, screen} from 'modules/testing-library';
+import type {DiagramOverlay} from './types';
 
-describe('AgentStatusOverlay', () => {
+const createOverlay = (status: AgentInstanceStatus): DiagramOverlay => ({
+  container: document.body,
+  elementId: 'Activity_agent',
+  type: 'agentStatus',
+  payload: {status, agentInstanceKey: 'agent-1'},
+});
+
+describe('agentStatusOverlay', () => {
   it.each<[AgentInstanceStatus, string]>([
     ['INITIALIZING', 'Starting...'],
     ['TOOL_DISCOVERY', 'Discovering tools...'],
@@ -19,7 +27,7 @@ describe('AgentStatusOverlay', () => {
   ])(
     'should render a status label for the active %s status',
     (status, label) => {
-      render(<AgentStatusOverlay container={document.body} status={status} />);
+      render(<AgentStatusOverlay overlay={createOverlay(status)} />);
 
       expect(
         screen.getByTestId(`agent-status-overlay-${status}`),
@@ -30,7 +38,7 @@ describe('AgentStatusOverlay', () => {
   it.each<AgentInstanceStatus>(['COMPLETED', 'IDLE'])(
     'should render nothing for the non-active %s status',
     (status) => {
-      render(<AgentStatusOverlay container={document.body} status={status} />);
+      render(<AgentStatusOverlay overlay={createOverlay(status)} />);
 
       expect(
         screen.queryByTestId(`agent-status-overlay-${status}`),
