+++ operate/client/src/modules/components/ModificationBadgeOverlay/styled.tsx
